change author id

+++ workspace/AcademicRetrieval/src/actm/data/base/BaseAuthor.java
@@ -30,7 +30,7 @@ public abstract class BaseAuthor  implements Serializable {
 	/**
 	 * Constructor for primary key
 	 */
-	public BaseAuthor (java.lang.Integer id) {
+	public BaseAuthor (java.lang.Long id) {
 		this.setId(id);
 		initialize();
 	}
@@ -39,7 +39,7 @@ public abstract class BaseAuthor  implements Serializable {
 	 * Constructor for required fields
 	 */
 	public BaseAuthor (
-		java.lang.Integer id,
+		java.lang.Long id,
 		java.lang.String name,
 		java.lang.String link,
 		java.lang.Integer index) {
@@ -58,7 +58,7 @@ public abstract class BaseAuthor  implements Serializable {
 	private int hashCode = Integer.MIN_VALUE;
 
 	// primary key
-	private java.lang.Integer id;
+	private java.lang.Long id;
 
 	// fields
 	private java.lang.String name;
@@ -77,7 +77,7 @@ public abstract class BaseAuthor  implements Serializable {
      *  generator-class="native"
      *  column="id"
      */
-	public java.lang.Integer getId () {
+	public java.lang.Long getId () {
 		return id;
 	}
 
@@ -85,7 +85,7 @@ public abstract class BaseAuthor  implements Serializable {
 	 * Set the unique identifier of this class
 	 * @param id the new ID
 	 */
-	public void setId (java.lang.Integer id) {
+	public void setId (java.lang.Long id) {
 		this.id = id;
 		this.hashCode = Integer.MIN_VALUE;
 	}
